refactor: api 폴더명 수정 [TRL-147]
